# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manajemen Proker API",
        "description": "Dokumentasi API untuk Sistem Manajemen Program Kerja. Mencakup Authentication, Menu Management, Role & Group Management, dan Permission Management.",
        "contact": {
            "email": "admin@manajemenproker.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/program-sie": {
            "get": {
                "tags": [
                    "Program Sie"
                ],
                "summary": "Daftar Sie Program (Paginated)",
                "description": "Mengambil daftar seluruh sie program dengan pagination.",
                "operationId": "1338abba42e084bcdf0118150cdc6967",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman aktif",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pagesize",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Program Sie"
                ],
                "summary": "Tambah Sie Baru",
                "description": "Membuat sie program baru beserta anggotanya.",
                "operationId": "5838c621b0898383e082c5fe8469e223",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_program",
                                    "sie_name"
                                ],
                                "properties": {
                                    "id_program": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sie_name": {
                                        "type": "string",
                                        "example": "Humas"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Sie Hubungan Masyarakat"
                                    },
                                    "id_koordinator": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "members": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id_user": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "Anggota"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/program-sie/{id}": {
            "get": {
                "tags": [
                    "Program Sie"
                ],
                "summary": "Detail Sie Program",
                "description": "Menampilkan detail satu sie program berdasarkan ID.",
                "operationId": "79e9a60c5ab4c97a00694134a0ef78be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Sie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Program Sie"
                ],
                "summary": "Update Sie Program",
                "description": "Memperbarui data sie program dan daftar anggotanya.",
                "operationId": "e20f0522c9d312aafff43b34c26856f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Sie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sie_name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "id_koordinator": {
                                        "type": "integer"
                                    },
                                    "members": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id_user": {
                                                    "type": "integer"
                                                },
                                                "role": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Program Sie"
                ],
                "summary": "Hapus Sie Program",
                "description": "Menghapus sie program berdasarkan ID.",
                "operationId": "3f11a97dcff81363a52ed89e3a5fa5ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Sie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dengan format: Bearer {access_token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}